# openapi/spec.yml
components:
  schemas:
...
    loginView:
      properties:
        email:
          format: email
          type: string
        id:
          format: uuid
          type: string
        password:
          format: password
          type: string
      type: object
...
    userCreate:
      properties:
        email:
          format: email
          type: string
        login_id:
          format: uuid
          type: string
        name:
          type: string
      required:
      - name
      - email
      - login_id
      type: object
    userPartial:
      properties:
        email:
          format: email
          type: string
        login_id:
          format: uuid
          type: string
        name:
          type: string
      type: object
...
    userUpdate:
      properties:
        email:
          format: email
          type: string
        login_id:
          format: uuid
          type: string
        name:
          type: string
      required:
      - name
      - email
      - login_id
      type: object
    userView:
      properties:
        email:
          format: email
          type: string
        id:
          format: uuid
          type: string
        login:
          $ref: '#/components/schemas/loginView'
        login_id:
          format: uuid
          type: string
        name:
          type: string
      type: object
  securitySchemes:
    bearer:
      scheme: bearer
      type: http
    cookieAccessToken:
      in: cookie
      name: x-access-token
      type: apiKey
    cookieRefreshToken:
      in: cookie
      name: x-refresh-token
      type: apiKey
    key:
      in: header
      name: X-API-Key
      type: apiKey
    oauth:
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          scopes:
            admin: Grants access to admin operations
            read: Grants read access
            write: Grants write access
          tokenUrl: https://example.com/oauth/token
      type: oauth2
info:
  description: Invoicing API Reference
  title: Invoicing API
  version: '1.0'
openapi: 3.0.3
paths:
...
  /api/v1/users:
    get:
      description: get all user entities
      operationId: getUsers
      parameters:
      - description: Filter Users by email
        in: query
        name: email
        schema:
          type: string
      - description: Filter Users by name
        in: query
        name: name
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/userView'
                type: array
          description: successful operation
      summary: get all user entities
      tags:
      - user
    post:
      description: Add a new user entity
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userCreate'
        description: Create a new user entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userView'
          description: successful operation
        '405':
          description: invalid input
      summary: Add a new user entity
      tags:
      - user
  /api/v1/users/{id}:
    delete:
      description: Delete an existing user entity
      operationId: deleteUser
      parameters:
      - description: user id to delete
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userView'
          description: successful operation
        '400':
          description: Invalid user id
        '404':
          description: user not found
      summary: Delete an existing user entity
      tags:
      - user
    get:
      description: Get an existing user entity
      operationId: getUser
      parameters:
      - description: user id to get
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userView'
          description: successful operation
        '400':
          description: Invalid user id
        '404':
          description: user not found
      summary: Get an existing user entity
      tags:
      - user
    patch:
      description: Modify an existing user entity (partial update, only provided properties
        will be updated)
      operationId: modifyUser
      parameters:
      - description: user id to partially update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userPartial'
        description: Modify an existing user entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: user not found
        '405':
          description: Invalid input
      summary: Modify an existing user entity
      tags:
      - user
    put:
      description: Update an existing user entity (all required properties must be
        supplied)
      operationId: updateUser
      parameters:
      - description: user id to update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userUpdate'
        description: Update an existing user entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: user not found
        '405':
          description: Invalid input
      summary: Update an existing user entity
      tags:
      - user
...
security:
- cookieAccessToken: []
  cookieRefreshToken: []
servers:
- url: http://localhost:8001
...